Get rid of frame writer
